backend/requirements.txt
Django
djangorestframework
Faker
wheel
